Adjust the syslinux recipe to only deploy an MBR when an MBR is created.

+++ recipes-bsp/syslinux/syslinux_6.01.bbappend
@@ -2,12 +2,15 @@
 PROVIDES += "virtual/bootloader"
 
 #
-#Deploy the bootloader (MBR) image to our images directory.
+#On BIOS systems, deploy the bootloader (MBR) image to our images directory.
 #
 do_deploy() {
     BOOTLOADER=${STAGING_DATADIR}/syslinux/mbr.bin
-    install -d ${DEPLOY_DIR_IMAGE}
-    install ${BOOTLOADER} ${DEPLOY_DIR_IMAGE}/syslinux-mbr-${MACHINE}.bin
-    install ${BOOTLOADER} ${DEPLOY_DIR_IMAGE}/bootloader-${MACHINE}.bin 
+
+    if [ -f $BOOTLOADER ]; then
+        install -d ${DEPLOY_DIR_IMAGE}
+        install ${BOOTLOADER} ${DEPLOY_DIR_IMAGE}/syslinux-mbr-${MACHINE}.bin
+        install ${BOOTLOADER} ${DEPLOY_DIR_IMAGE}/bootloader-${MACHINE}.bin
+    fi
 }
 addtask deploy before build after populate_sysroot
